refactor(backend): amount Decimal(15,2) → BigInteger rubles (БАГ-5 part 1/2)
Models, schemas, services, endpoints, alembic migration. Partitioned
t_f_budget_fact: declarative range-partitioning propagates ALTER
automatically. Validators переписаны на gt=0/strict=True.
ROUND(amount)::BIGINT в migration: фракционных amount в проде нет.

Frontend, bot, и Dexie — следующим коммитом.

+++ backend/app/services/notification_service.py
@@ -6,7 +6,6 @@ Handles automated notifications:
 - Budget threshold checks (FR-006)
 """
 from datetime import date, timedelta
-from decimal import Decimal
 
 import httpx
 from sqlmodel import select
@@ -74,8 +73,8 @@ class NotificationService:
         article_id: int,
         notification_type: str,
         threshold_percent: int,
-        plan_amount: Decimal,
-        actual_amount: Decimal,
+        plan_amount: int,
+        actual_amount: int,
         period_start: date,
         period_end: date,
         user_id: int | None = None,
